Fix: Cambiar tipos de parámetros de int a float en métodos de stock
- Los parámetros $nuevaReservada y $nuevaDisponible estaban como int&
- Causaba truncamiento de decimales: 1.95 → 1
- Cambiar todos los métodos de aplicación a float&
- Esto arregla el problema de 'stock inconsistente' con productos en kg

+++ app/Services/Stock/MovimientoStockService.php
@@ -457,8 +457,8 @@ class MovimientoStockService
     private function aplicarReservaProforma(
         float $cantidad,
         float &$nuevoTotal,
-        int &$nuevaReservada,
-        int &$nuevaDisponible
+        float &$nuevaReservada,
+        float &$nuevaDisponible
     ): void {
         $cantidadAReservar = abs($cantidad);
         $nuevaReservada += $cantidadAReservar;
